Impl clan tournament system

+++ src/main/java/me/Vark123/EpicClans/ClanSystem/ClanTreasury.java
@@ -16,21 +16,21 @@ public class ClanTreasury {
 	private int pr;
 	
 	public void addMoney(double money) {this.money += money;}
-	public void addStygia(double stygia) {this.stygia += stygia;}
-	public void addCoins(double coins) {this.coins += coins;}
-	public void addRuda(double ruda) {this.ruda += ruda;}
-	public void addPr(double pr) {this.pr += pr;}
+	public void addStygia(int stygia) {this.stygia += stygia;}
+	public void addCoins(int coins) {this.coins += coins;}
+	public void addRuda(int ruda) {this.ruda += ruda;}
+	public void addPr(int pr) {this.pr += pr;}
 	
 	public void removeMoney(double money) {this.money -= money;}
-	public void removeStygia(double stygia) {this.stygia -= stygia;}
-	public void removeCoins(double coins) {this.coins -= coins;}
-	public void removeRuda(double ruda) {this.ruda -= ruda;}
-	public void removePr(double pr) {this.pr -= pr;}
+	public void removeStygia(int stygia) {this.stygia -= stygia;}
+	public void removeCoins(int coins) {this.coins -= coins;}
+	public void removeRuda(int ruda) {this.ruda -= ruda;}
+	public void removePr(int pr) {this.pr -= pr;}
 	
 	public boolean hasEnoughMoney(double money) {return this.money >= money;}
-	public boolean hasEnoughStygia(double stygia) {return this.stygia >= stygia;}
-	public boolean hasEnoughCoins(double coins) {return this.coins >= coins;}
-	public boolean hasEnoughRuda(double ruda) {return this.ruda >= ruda;}
-	public boolean hasEnoughPr(double pr) {return this.pr >= pr;}
+	public boolean hasEnoughStygia(int stygia) {return this.stygia >= stygia;}
+	public boolean hasEnoughCoins(int coins) {return this.coins >= coins;}
+	public boolean hasEnoughRuda(int ruda) {return this.ruda >= ruda;}
+	public boolean hasEnoughPr(int pr) {return this.pr >= pr;}
 	
 }
